Updated with StyleCop docs, last before Plugin
Fleshed out all the documentation.  No functional changes made, but a
couple minor tweaks were done while documenting.  This commit
corresponds to v1.1.1, but it will not be publicly released.  This is
the last commit before changing this to be a Plugin which is a new
feature of the OCTGN deck editor.